Refactor project structure/scaffolding
- All the src code is in src/ instead of api/ now. package.json scripts reflect this.
- config/secrets.js is a single file that serves the same function. We don't really need a directory for this as it's always going to be just 1 file.
- Stuffed a bunch of misecllaneous stuff into a folder called utils. The interestData should probably be a seed file.
- Moved all the routes into routes.js. This needs heavy refactoring and testing but I didn't want to do all that in this PR. Plus, if I messed with the route strings, it would probably break the react-app.
