--- config/secrets.js
@@ -1,9 +1,0 @@
-const port = process.env.PORT || "8080"
-const seedPWTest = process.env.SEED_PW_TEST
-const secret = process.env.JWT_SECRET || "sdfgsdfgsd~!@`21123asdf~!@#"
-
-module.exports = {
-    port,
-    seedPWTest,
-    secret
-}
