+++ package.json
@@ -2,7 +2,7 @@
     "name": "friendlier_backend",
     "version": "0.1.0",
     "private": true,
-    "main": "./api/server.js",
+    "main": "src",
     "dependencies": {
         "bcryptjs": "^2.4.3",
         "body-parser": "^1.19.0",
