+++ src/config.js
@@ -1,0 +1,9 @@
+const {
+    PORT = "8080",
+    JWT_SECRET = "sdfgsdfgsd~!@`21123asdf~!@#"
+} = process.env
+
+module.exports = {
+    PORT,
+    JWT_SECRET
+}
